settings: UI reorg, per-library update, Who's watching, transcoding/sync tweaks
- Reorder tabs: Account, Library, Sync, Playback, Transcoding.
- Drop the deviceName setting; report Kodi's System.FriendlyName (as pvr.kofin).
- Blank group-label dividers where they reclaim vertical space; remove the
  Account "Connection" separator (sslVerify merged up).
- Force direct play keeps preferred codecs + transcode audio bitrate visible.
  The direct-play codec/HDR lists stay visible too: hiding them would need a
  settings-window close, and the Omega list[string] dependency bug rules out
  native hiding.
- Remove E-AC3 from the direct-play audio codecs.
- Context transcode gains a Source option (0) and 0.5/0.75 Mbit/s, with the
  bitrate plumbed as a float through play -> deviceprofile.
- Per-library Update picker (mirrors Repair); tooltips note when each action
  runs. "Update all libraries" -> "Update libraries".
- Media-segment defaults: Commercials Auto, the rest Ask; Play Next off.
- Remove sync toast notifications; add a syncProgressThreshold slider that
  hides Kodi's library-update progress below N items (default 20).
- "Add user to session" -> "Who's watching?": one multiselect that toggles
  additional users (the permanent logged-in user shown in the title); the root
  menu label refreshes to list them.

Verified live on Kodi 21.3: force-DP visibility, the Who's-watching dialog, and
the reordered/blank-divider layout.

+++ lib/kofin/plugin/play.py
@@ -149,9 +149,9 @@ def play(request: Request) -> None:
 
     transcode = request.params.get("transcode") == "1"
     try:
-        bitrate_mbps = int(request.params.get("bitrate", "0"))
+        bitrate_mbps = float(request.params.get("bitrate", "0"))
     except ValueError:
-        bitrate_mbps = 0
+        bitrate_mbps = 0.0
 
     api = Api.from_credentials(Http(settings.get_bool("sslVerify")), creds)
     try:
